fix: Change strategy, use resource still and move cluster model to epiniomgmt directory:
